Moved QC class to weewx/qc.py
